pyproject.toml
[tool.black]
line-length = 120
target-version = ['py39']
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | \.env
  | \.idea
  | build
  | dist
  | cdk.out
)/
'''

[tool.isort]
profile = 'black'
line_length = 88
force_single_line = true

[tool.poetry]
name = "BBDN-Sign-Up-List-Tool"
version = "0.1.0"
description = "Example of a LTI app using Learn REST API"
authors = []

[tool.poetry.dependencies]
python = "^3.9.0"
Flask-RESTful = "^0.3.9"
Flask-Caching = "^1.10.1"
boto3 = "^1.21.22"
aws-lambda-wsgi = "^0.0.6"
Flask = "^2.0.3"
python-dotenv = "^0.19.2"
python-decouple = "^3.6"
requests = "2.27.0"
importlib-metadata = "^4.11.3"
cachetools = "^5.0.0"
pydantic = "^1.9.1"
PyJWT = {extras = ["crypto"], version = "^2.4.0"}
zipp = "^3.8.0"
cffi = "^1.15.0"
jwcrypto = "^1.3.1"

[tool.poetry.dev-dependencies]
"aws-cdk-lib" = "^2.0.0"
"constructs" = "^10.0.0"
toml = "^0.10.2"
pytest = "^6.2.4"
SQLAlchemy = "^1.4.32"
gunicorn = "^20.1.0"
pre-commit = "^2.19.0"
selenium = "^4.1.3"
screenpy-selenium = "^4.0.2"
pytest-bdd = "^5.0.0"
assertpy = "^1.1"
moto = "^3.1.13"
pytest-mock = "^3.7.0"
exceptiongroup = "^1.0.0-rc.8"
autopep8 = "^1.6.0"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
